requirements.txt
antiorm==1.2.1
astroid==2.5.1
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.15.3
pylint==2.7.2
PyYAML==5.4.1
simpy==4.0.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
PyQt5~=5.15.4